@dougwilson 2015-09-01
I have tweaked the style in the commit, so please double-check everything :)