@TimothyGu 2015-09-01
@dougwilson found a minor incompatibility: https://github.com/component/escape-html/commit/da36164db117468d06118e09f51676237466cb33#commitcomment-12989394