@dougwilson 2015-09-01
Sorry, didn't notice the `String(html)` in the original; I'll rewrite history. The reason I removed it was because you were altering an argument, which is slow.