@dougwilson 2015-09-01
Looks like that took a hit on the performance ;)