@TimothyGu 2015-09-01
> Looks like that took a hit on the performance ;)

Try `'' + str` rather than `String(str)`; it's a lot faster (and shorter) here on both Node.js 0.10.38 and io.js 3.2.0.