@dougwilson 2015-09-01
Hm. There must be some reason for the difference. Do both result in the exact same behavior? I'm trying to look up the ECMA 5 spec for this.