@dougwilson 2015-09-01
Changing this may also be technically non-backwards compatible for people who have replaced the global `String` function ;)