@dougwilson 2015-09-01
Comparing http://www.ecma-international.org/ecma-262/5.1/#sec-15.5.1.1 and http://www.ecma-international.org/ecma-262/5.1/#sec-11.6.1 shows there is no functional difference between the two :)