@TimothyGu 2015-09-01
See #18, seems like you beat me to it when it comes to ES spec reading :)

But #18 also has an additional perf fix.